Less harsh language around travel expenses
At least in the US there are hundreds of free community events with very limited budgets. Most of them are non-profits and the organizers are all volunteers. The language around "the minimum you can do" is unrealistic for many of those events.

I think it's better to be clear about what you can and can't do. If you can't afford to pay for speakers to travel in, that's ok, but realize you're only going to get local speakers.